pyproject.toml
[project]
name = "app-report-devops-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "httpx>=0.28.1",
    "motor>=3.7.1",
    "pydantic-settings>=2.10.1",
    "pymongo>=4.14.1",
]
